+++ lib/delayed/performable_method.rb
@@ -1,10 +1,6 @@
-require 'active_support/core_ext/module/delegation'
-
 module Delayed
   class PerformableMethod
     attr_accessor :object, :method_name, :args
-
-    delegate :method, :to => :object
 
     def initialize(object, method_name, args)
       raise NoMethodError, "undefined method `#{method_name}' for #{object.inspect}" unless object.respond_to?(method_name, true)
@@ -30,9 +26,15 @@ module Delayed
       object.send(method_name, *args) if object
     end
 
+    def method(sym)
+      object.method(sym)
+    end
+
+    # rubocop:disable MethodMissing
     def method_missing(symbol, *args)
       object.send(symbol, *args)
     end
+    # rubocop:enable MethodMissing
 
     def respond_to?(symbol, include_private = false)
       super || object.respond_to?(symbol, include_private)