+++ spec/autoloaded/struct.rb
@@ -1,7 +1,7 @@
 # Make sure this file does not get required manually
 module Autoloaded
-  class Struct < ::Struct.new(nil)
-    def perform
-    end
+  Struct = ::Struct.new(nil)
+  class Struct
+    def perform; end
   end
 end