+++ spec/performable_method_spec.rb
@@ -30,8 +30,7 @@ describe Delayed::PerformableMethod do
 
   it 'does not raise NoMethodError if target method is private' do
     clazz = Class.new do
-      def private_method
-      end
+      def private_method; end
       private :private_method
     end
     expect { Delayed::PerformableMethod.new(clazz.new, :private_method, []) }.not_to raise_error
@@ -68,7 +67,7 @@ describe Delayed::PerformableMethod do
       story = Story.create
       expect(story).to receive(:error).with(an_instance_of(Delayed::Job), an_instance_of(RuntimeError))
       expect(story).to receive(:tell).and_raise(RuntimeError)
-      expect { story.delay.tell.invoke_job }.to raise_error
+      expect { story.delay.tell.invoke_job }.to raise_error(RuntimeError)
     end
 
     it 'delegates failure hook to object' do
@@ -98,7 +97,7 @@ describe Delayed::PerformableMethod do
         story = Story.create
         expect(story).to receive(:error).with(an_instance_of(Delayed::Job), an_instance_of(RuntimeError))
         expect(story).to receive(:tell).and_raise(RuntimeError)
-        expect { story.delay.tell }.to raise_error
+        expect { story.delay.tell }.to raise_error(RuntimeError)
       end
 
       it 'delegates failure hook to object' do