+++ spec/worker_spec.rb
@@ -24,12 +24,20 @@ describe Delayed::Worker do
   describe 'job_say' do
     before do
       @worker = Delayed::Worker.new
-      @job = double('job', :id => 123, :name => 'ExampleJob')
+      @job = double('job', :id => 123, :name => 'ExampleJob', :queue => nil)
     end
 
     it 'logs with job name and id' do
+      expect(@job).to receive(:queue)
       expect(@worker).to receive(:say).
         with('Job ExampleJob (id=123) message', Delayed::Worker.default_log_level)
+      @worker.job_say(@job, 'message')
+    end
+
+    it 'logs with job name, queue and id' do
+      expect(@job).to receive(:queue).and_return('test')
+      expect(@worker).to receive(:say).
+        with('Job ExampleJob (id=123) (queue=test) message', Delayed::Worker.default_log_level)
       @worker.job_say(@job, 'message')
     end
 
@@ -152,6 +160,24 @@ describe Delayed::Worker do
       expect(@worker.logger).to receive(:send).
         with('info', "#{@expected_time}: #{@worker_name} #{@text}")
       @worker.say(@text, Delayed::Worker.default_log_level)
+    end
+  end
+
+  describe 'plugin registration' do
+    it 'does not double-register plugins on worker instantiation' do
+      performances = 0
+      plugin = Class.new(Delayed::Plugin) do
+        callbacks do |lifecycle|
+          lifecycle.before(:enqueue) { performances += 1 }
+        end
+      end
+      Delayed::Worker.plugins << plugin
+
+      Delayed::Worker.new
+      Delayed::Worker.new
+      Delayed::Worker.lifecycle.run_callbacks(:enqueue, nil) {}
+
+      expect(performances).to eq(1)
     end
   end
 end