+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+4.1.5 - 2018-04-13
+=================
+* Allow Rails 5.2
+
+4.1.4 - 2017-12-29
+=================
+* Use `yaml_tag` instead of deprecated `yaml_as` (#996)
+* Support ruby 2.5.0
+
+4.1.3 - 2017-05-26
+=================
+* Don't mutate the options hash (#877)
+* Log an error message when a deserialization error occurs (#894)
+* Adding the queue name to the log output (#917)
+* Don't include ClassMethods with MessageSending (#924)
+* Fix YAML deserialization error if original object is soft-deleted (#947)
+* Add support for Rails 5.1 (#982)
+
+4.1.2 - 2016-05-16
+==================
+* Added Delayed::Worker.queue_attributes
+* Limit what we require in ActiveSupport
+* Fix pid file creation when there is no tmp directory
+* Rails 5 support
+
+4.1.1 - 2015-09-24
+==================
+* Fix shared specs for back-ends that reload objects
+
+4.1.0 - 2015-09-22
+==================
+* Alter `Delayed::Command` to work with or without Rails
+* Allow `Delayed::Worker.delay_jobs` configuration to be a proc
+* Add ability to set destroy failed jobs on a per job basis
+* Make `Delayed::Worker.new` idempotent
+* Set quiet from the environment
+* Rescue `Exception` instead of `StandardError` in worker
+* Fix worker crash on serialization error
+
 4.0.6 - 2014-12-22
 ==================
 * Revert removing test files from the gem